unibind: share one staged eventually helper across ex suites
The eventually/2 + poll/2 polling helper was copied verbatim into the
plumb, tui, and conformance binding suites (the 26-line type1 clone
group the Elixir gate revealed). The ex target now stages a single
packages/unibind/nix/ex-support/eventually.exs into every assembled mix
package as test/support/eventually.exs; the suites import it and the
copies are deleted. Ratchet tightens 0.37 -> 0.365 (measured 0.3623).

See packages/site/src/lib/updates/unibind-ex-shared-eventually.svx.

Fixes #3901
