+++ packages/unibind/nix/ex.nix
@@ -95,9 +95,16 @@
       ${findCdylib}
       install -m555 "$cdylib" "$out/priv/native/${soname}"
       ${lib.optionalString (mixSource != null) ''
+        # Shared ExUnit support for every binding suite (loaded from the
+        # suite's test_helper.exs): NIF-side observables are asynchronous,
+        # so suites poll via one staged helper instead of per-suite copies
+        # (#3901). Staged ahead of the overlay loop so a hand-written copy
+        # trips the same shadow check as a generated file.
+        mkdir -p "$out/test/support"
+        cp ${./ex-support/eventually.exs} "$out/test/support/eventually.exs"
         (cd ${mixSource} && find . -type f -print) | sed 's|^\./||' | while read -r file; do
           if [ -e "$out/$file" ]; then
-            echo "unibind: $file is generated by unibind-gen; delete the hand-written copy" >&2
+            echo "unibind: $file is generated or staged by the ex target; delete the hand-written copy" >&2
             exit 1
           fi
           mkdir -p "$out/$(dirname "$file")"
